requirements.txt
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-llms-nvidia
llama-index-embeddings-nvidia
llama-index-llms-groq
llama-index-llms-gemini
llama-index-utils-workflow
llama-parse
python-dotenv
streamlit
torch